Remote Resolution of Tasks (#97)
Prior to this change, the project used Tasks applied directly to
the cluster. In this change, we switch to [remote resolution] of
Tasks in a Pipeline. This is aligned with the  "config as code"
recommendation, where build definitions are fetched from a version
control system.

In addition, we make a few minor changes:
- cleanup the documentation
- remove unused result from the upload-model task

[remote resolution]: https://tekton.dev/docs/pipelines/resolution/