Add support for hashing files with header.
Missed this in #188, but found out I need it when working on #190. The
`serialize_v0`/`serialize_v1` methods all had headers in front of the
files, so we need to do that too. Will update usage of header on #190
shortly.

As a benefit, we can simulate hashing a file with a header for the first
portion of the file and a sharded hasher for the remainder of the file.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>