Lint:test for missing endline at eof, fix errors.
We were missing a test that all text files end with a newline character.

Also, the old `tac | awk` process to count the number of empty lines at
the end of a file was prone to failures if `awk` exits early. So I
replaced it with a test that should be more resilient.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>