Add support for `hashlib.file_digest`.
This should hash files without going via Python's I/O, but requires
caller to open the file descriptor in advance and close it afterwards.

For now, rather than putting the file descriptor open/close inside the
class, we're leaving it to the caller, as this could be used to benefit
while serializing a model, or while reading the model from a cloud
filesystem (I think).

I needed to do several renames to make sure API stays the same. I also
fixed a test that was testing the wrong class.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>