Reduce fixture scope to class instead of session. (#242)
Session scope means the fixtures are created at the start of `pytest` run and finalized when the testing ends. This means their lifetime is much longer than what we need.

Class scope means that fixtures are created when the test class is first initialized and finalized when all the tests under that class have finished. This is more appropriate to what we want to use these fixtures for.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>