Use `conftest.py` instead of `fixtures.py`.
The way to properly configure `pytest` is via `conftest.py`. However, internal guidance recommended to use a plugin-like approach instead, to help with discoverability. But, this does not work if we want to add command line arguments to the tests (which we need to add goldens support so we don't have to manually type long hashes as we change code). But, the guidance only applies to google-only code, not to OSS, so we can switch to `conftest.py`.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>