Migrate tests for model serialization to goldens. (#244)
This is to ensure that whenever we change the model serialization code, it is much easier to update the tests. The changes should be more readable this way.

When code changes to request changes to the goldens, the author of the change will need to run a `pytest` command to update the files by passing a flag, then add the goldens to the PR (ideally via a new commit).

This also ensures that we can test all models in all scenarios, in a similar way. Thus, removed the tests for empty models that were just testing for empty manifest. Added a new model that only contains an empty file.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>