Move test support to root of model_signing.
Since we will need to share some of the models with other directories
besides serialization, the only way to achieve this without duplication
is to have `conftest.py` in the root testing directory. And, when moving
that, we also need to move `test_support.py` which defines useful
constants used both in tests and in `conftest.py`.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>