combine model serialization logic for files and directories
Identical logic was being performed whether model_path pointed to a file
or a directory due to two pathlib.Path.glob behaviors:

1. The path must be a directory for the generator to yield items
2. The glob does not include the provided directory itself.

By iterating over the provided model_path first, and then continuing to
the glob, all of the serialization logic can be done in the loop. This
will help future changes which will add more serialization logic in the
form of ignore paths.

Signed-off-by: Spencer Schrock <sschrock@google.com>