Add machinery for model signing and verification.
Added a set of `Empty*` classes just to show how these would be used in OSS. I'll send another PR for the actual in-toto classes, but this one mirrors the internal changelist where the API gets introduced.

For the 2 `serialize_*_test.py` files: just ordered the imports to match the internal style.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>