Expose resource descriptors from manifests (#256)
* Expose reosurce descriptors from manifests

We aim this to be similar to in-toto's `ResourceDescriptor`. To support cases where in-toto cannot be directly used, we make this a dataclass that can be mapped to in-toto when needed, and used as its own otherwise.

Not all fields from in-toto are specified at this moment. All fields here must be present, unlike in-toto, where all are optional.

See https://github.com/in-toto/attestation/blob/main/spec/v1/resource_descriptor.md for the in-toto specification.

This is the first separable PR for the signing support (see full draft on #253)

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Remove f-strings where not needed

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Handle internal review

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Remove "."

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>