Add in-toto format for items with hash of hashes.
This converts model serialization manifests that record every model file
hash into an in-toto payload that can then be passed to Sigstore's
`sign_intoto` for signing to generate a Sigstore `Bundle` (if using Sigstore).

To identify the models, we compute a hash of all hashes of the files and
use that as the subject. The individual file hashes are used as the
payload and we would have the verifier check them as part of the
verification process.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>