Add in-toto format with hashes of files as subjects
This converts model serialization manifests that record every model file
hash into an in-toto payload that can then be passed to Sigstore's
`sign_intoto` for signing to generate a Sigstore `Bundle` (if using
Sigstore).

This time, we record every hash as part of the subject instead of in the
payload. We require verifiers to be aware of this and acknowledge that
verifiers that only check subject by subject (that is, they check if the
hash of a passed in argument is in the list of subjects and don't check
if all the hashes are present), can fail to fully detect if the model
integrity is compromised by renaming one file in the model,
interchanging two file names, or deleting a file. The signing library
will have additional checks for this, but verifying the signature with
other tools might result in invalid results.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>