Add a `from_str` method to `manifest.Shard`.
The `ShardedFileManifestItem` objects (items in `ShardLevelManifest`)
are recorded as a single string in the in-toto payloads used for
signing. The canonicalization to string is done by appending the file
(relative) path, the start offset and the end offset, separated by `:`.

When validating a signature and rebuilding a manifest, we need to parse
a string back to a shard. Rather than replicating this in all places, we
create a member function for the conversion.

Also adds the relevant tests.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>