Reorganize code to enable wheel publishing
This commits moves files around to get to a `src/` layout
(https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/#src-layout-vs-flat-layout)
for the code. There are more benefits for a `src/` layout than a flat
layout, such as:

- documentation is easier to generate
- we won't accidentally test against code that is different than what
  users would see
- we can include only the source in the wheel, reducing PyPI bandwidth
- it makes it clearer which parts of the repo go into the wheel and
  which ones are ancillary. Incidentally, we might now want to move the
  SLSA parts to another repo and maybe rename this one? We can do this
  at a later time.

For this commit I only moved the files around. The only change is in
`README.md` where I replaced the pointer to `model_signing/README.md`
with one to `README.model_signing.md`, as I'm creating this file now.

Since we would need to change `README.model_signing.md` anyway once the
API is ready, I have not updated it besides just fixing the image
pointer.

I have deleted old files from when this was a POC. Since we're working
on releasing the library, we no longer need those.

The alternative to not moving the files to the root of the folder would
have been to duplicate LICENSE, etc. inside the `model-signing`
directory, so that packaging tools would see it. I decided against that,
since we still needed to move files around to make a `src/` directory
and it did no longer make sense to also do file duplication.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>