Update GHAs to use hatch for testing and linting
The scripts are now much simpler and easier to update to newer versions
of Python.

I also added copyright notices to all files. Removed the workflows that
are no longer relevant (slsa_for_ml needs to be pinned right now to
first make sure it works -- probably we'll move it to a separate repo).

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>