Add unit tests for signing with sigstore (#291)
* Configure coverage reporting

By default, `hatch test -c` coverage report lists only counts of lines and lines missing and percentages but there is no way to see which are the ones that are missing. We don't have an option to generate an html report at the moment (https://github.com/pypa/hatch/pull/1477). Added some options to display missing lines.

Also, by default, all files are included in the report, including tests (covering the test-only code). I removed the tests, but if we decide we should add them that's easy to do.

More importantly, the report lists files that are 100% covered (not useful in CI) and empty files (not useful at all). So, I removed those from the output.

There is another bigger issue that only files that are imported by a test get reported, so if we have code that is not tested at all it will not show up here. We already have such code in `signature/` and `signing/sigstore.py`. Fixing this will be left for later.

Current output is:

```
Name                                         Stmts   Miss  Cover   Missing
--------------------------------------------------------------------------
src/model_signing/signing/in_toto.py           168     68    60%   65-78, 181-190, 342-367, 485-512, 660-671, 793-806
--------------------------------------------------------------------------
TOTAL                                          745     70    91%
```

Fixing the missing coverage is left for later. We should aim for 95%+ or so coverage, I think.

We should probably make it so that GitHub reports this table back on PRs, so reviewers can quickly ask for more testing without needing to check the GHA report. Punted for later, for now I'll just remember to just keep checking.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add unit tests for signing with sigstore.

We need to do quite a lot of mocking around Sigstore, but we are able to test all logic in our library. What is left to do for testing is e2e integration tests (#5) and testing with signing on one OS and verifying on another (#25). Both of these are integration style tests and we will only be able to run them in GHA. I'll send a PR for those soon.

While testing, I discovered some minor bugs with error reporting and one moderate bug. Fixed in this PR.

We now have achieved 100% test coverage! :tada:

```
Name    Stmts   Miss  Cover   Missing
-------------------------------------
TOTAL     835      0   100%
```

Well, almost. There are 2 files that are not imported by tests at all, so they don't get included in the report:

```
src/model_signing/signature/fake.py
src/model_signing/signature/pki.py
```

This depends on #287 which configures the coverage reporting.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>