Add a higher level API.
Currently, this only supports Sigstore (and only DSSE on verification).
I want to push this now so that people can start playing with it while I
update the Colab demo and try to make the API more uniform and add
support for the missing cases.

With this, the default signing should be as simple as

```
from model_signing import api as model_signing_api

model_signing_api.sign(model_path, signature_path)
```

And verification would be as simple as:

```
from model_signing import api as model_signing_api

model_signing_api.verify(
  model_path, signature_path,
  expected_identity, expected_oidc_provider
)
```

Will think on how to simplify this even further..

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>