change default shard size to 1 GB
The values was based on the benchmarks in b2ddcc0.

The exact speed improvement depends on the size of the models being
serialized and of the individual files in the model. Speed improvements
ranged from a 5% improvement to a 87% improvement.

Manifest size is also influenced by shard size (and the model /
individual file size). Increasing shard size from ~1MB to 1GB decreases
manifest size by 3 orders of magnitude (99.9% reduction).

Signed-off-by: Spencer Schrock <sschrock@google.com>