change default shard size to 1GB (#357)
* change default shard size to 1 GB

The values was based on the benchmarks in b2ddcc0.

The exact speed improvement depends on the size of the models being
serialized and of the individual files in the model. Speed improvements
ranged from a 5% improvement to a 87% improvement.

Manifest size is also influenced by shard size (and the model /
individual file size). Increasing shard size from ~1MB to 1GB decreases
manifest size by 3 orders of magnitude (99.9% reduction).

Signed-off-by: Spencer Schrock <sschrock@google.com>

* update shard test goldens

Signed-off-by: Spencer Schrock <sschrock@google.com>

* use gigabyte instead of gibibyte

They're roughly equal in performance, any differences either vary model
to model or run to run. However multiples of 1000 are slightly easier
for humans to visualize in things like shard names.

Signed-off-by: Spencer Schrock <sschrock@google.com>

---------

Signed-off-by: Spencer Schrock <sschrock@google.com>