remove _FileDigestTree from serialize_by_file.DigestSerializer
Standardizes the DigestSerializer merge behavior to match what's done in
the `serialize_by_file_shard` equivalent. Particularly, the file hasher
was turned into a factory, so the hashing can be done in parallel via
max_workers.

Merging the digests into a final digest now only looks at files, and not
the directories as well.

Signed-off-by: Spencer Schrock <sschrock@google.com>