cli: Add support for ignoring paths when signing
Add support for ignoring paths that should not be part of the signature.
This is useful when for example a model is signed before it is checked
into a git repository where files like .gitattributes or a .git directory
may be added (see models in huggingface git repos). In this case a
verifier should be made aware that git-related files are to be ignored.

The signing with a private EC key can be achieve like this:

python -m model_signing sign key \
   --signature /path/to/model/model.sig \
   --ignore-paths /path/to/model/.git \
   --ignore-paths /path/to/model/.gitattributes \
   --private-key mysigningkey.pem \
   /path/to/model

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>