cli: Add support for ignoring paths when signing or verifying (#373)
* cli: Add support for ignoring paths when verifying

Add support for ignoring paths that should not be part of the signature
verification. This is useful when for example a model was signed before
it was check into a git repository where files like .gitattributes or a
.git directory may have been added (see models in huggingface git repos).
In this case a verifier needs to be made aware that git-related files
are to be ignored and signature verification may then be achieved like
this:

python -m model_signing verify sigstore \
   --signature /path/to/model/model.sig \
   --identity <identity> \
   --identity_provider <identity provider> \
   --ignore-paths /path/to/model/.git \
   --ignore-paths /path/to/model/.gitattributes \
   /path/to/model

Example: https://huggingface.co/deepseek-ai/DeepSeek-V3-0324/tree/main

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* cli: Add support for ignoring paths when signing

Add support for ignoring paths that should not be part of the signature.
This is useful when for example a model is signed before it is checked
into a git repository where files like .gitattributes or a .git directory
may be added (see models in huggingface git repos). In this case a
verifier should be made aware that git-related files are to be ignored.

The signing with a private EC key can be achieve like this:

python -m model_signing sign key \
   --signature /path/to/model/model.sig \
   --ignore-paths /path/to/model/.git \
   --ignore-paths /path/to/model/.gitattributes \
   --private-key mysigningkey.pem \
   /path/to/model

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* cli: Add support for --ignore-git-paths option

Implement support for --ignore-git-paths to automatically ignore git-
related files in the MODEL_PATH. It is set to 'True' by default and
therefore the following now works and automatically ignores git-related
files.

python -m model_signing verify key \
     --signature /path/to/model/model.sig \
     --public_key ./public.pem \
     /path/to/model

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>