Remove artifact signing, keep only DDSE/in-toto.
This is cleanup to reach a single signature format to ease adoption.

Artifact signing requires knowing the hash in advance before
verification and this causes API mismatches. CLI only uses DSSE/in-toto,
and it doesn't make sense to support it only for API, especially since
we did not find any users for this. Plus, computing a single model hash
for this signing is much much slower.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>