Make hashing private and with consistent comments. (#388)
This is first part in controlling what the public API is.

We are still making sure the comments are comprehensive to help in IDEs.

I changed the name of the hashers. For the file hasher, since the result
is the same as reading the file and then hashing the contents, I changed
the name to just use the content hasher's name (so instead of
`file-sha256` it would be `sha256`). This resolves some inconsistencies
in serialization goldens.

For the sharded hasher, rather than start with `file` the name is now
`<algorithm>-sharded-<shard_size>`, which is more informative.

I removed the opened file descriptor hash engine for now. We have not
found a use for it, we needed a polyfill to handle Python versions below
3.11.

Had to fix a test: the empty files and directories are now ignored so a
model that is an empty file should hash exactly the same as a model that
is an empty directory.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>