Simplify manifest interface (#391)
* Simplify manifest interface

I'm doing a lot of items here, I can split them into multiple PRs if needed.

First, I'm making the `manifest/manifest.py` file be just `_manifest.py`. Removes an `__init__.py` and one level of import and marks this API as private too. For releases past v1, when we want to offer ability to partially match against a manifest we can either open up this API or just provide helper functions.

Next, I'm removing `DigestManifest`. It does not make sense to maintain the serializers that compute a single hash for the entire model blob. They are slow. For the cases where `DigestManifest` was used to just wrap around a digest, we can just use the `Digest` object directly.

This removal also means removal of the obsolete `DigestSerializer`s. These are not used since we want to record every component of the model with their hash instead of just one hash. This removal opens up other opportunities to remove dead code from the serialization files, but that would come as a separate PR.

Furthermore, we can remove `SingleDigestIntotoPayload`, as it is no longer needed, and was never used for actual signatures. There are further cleanup opportunities in `signing/...`, left for future PRs.

Next, since now all manifests are a collection of items, rather than have two separate classes we switch to having `Manifest` contain everything and then have classes for the keys and values that can be held inside the manifest.

Finally, I'm making sure that all comments in the manifest file are correct, and consistent.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add some casts to fix typechecking in tests

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Rename test file to match src

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>