Reduce verbosity in file hashing (#395)
Based on a quick idea, instead of `file_hashing.*` we can do `io.*`.
Now, we no longer have duplication (`file_hashing.FileHasher`) and the
import is much shorter, we reduce a few of the function calls that are
on 3 lines.

Plus, if we later add back hashing of cloud filesystems (#148), the name is
better suited.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>