Remove unused intoto formats. (#398)
Keep only `DigestsIntotoPayload` for now, but in a future PR this will also get replaced. Goal is to have one single intoto format with ability to restore hashing information (#331). This will also allow extending the signature to contain additional information in the future.

Right now it looks like there is no way to distinguish between file serialization and shard serialization. This is because the CLI was only testing file serialization, but future PR will make it work with both.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>