Fold payload classes up the class hierarchy (#399)
#### Summary
After #398 we have only one abstract class for the payload but with a subclass which then has another subclass. To simplify everything and prepare for #331, I'm folding everything into one single type now. Every step of the folding is one individual commit.

I'll change the names to remove some repetition once #331 is done.

I removed some type checks which are no longer needed, but not all of them. Fixed some of the comments, but not all of them. Future work is needed, after #331, likely.

* Fold `DigestsIntotoPayload` into `IntotoPayload`.

We have one abstract class with just one implementation after the removal in #398.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Fold `IntotoPayload` into `SigningPayload`.

We want to support only one signing payload type.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>