Change to new format for the signing payload (#407)
* Change docstring of the payload class.

This is to document the changes that would come in the next commits. The
new payload format is included. There are no other changes except the
docstring here.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Change `predicate_type` to match new format

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add the remaining changes to the payload docstrings

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add model name to manifest, update serialization

The serializers will now need to take the last entry in the model path
and use that as the model name. This is purely informative, but it's
better to have e.g. "gemma" in the signature than ".".

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Change payload to record model name and root hash

On the verification path, the `items` list is kept empty as we still
cannot know if it is a list of files or a list of shards. But the root
hash computation and the model name should properly roundtrip.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Record serialization method in manifest & payload

This ensures that we can correctly deserialize a payload into a
manifest.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Update `manifest_test` with the new `Manifest`.

The constructor now has 3 arguments, so we need to properly update here.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Refactor golden testing to separate method.

We'll add more similar goldens here, so extract out the most common
method.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add test for sharded models (two sizes)

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add all updated goldens

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Handle review

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>