Add traditional signing signers and verifyiers.
First, we add a new `sign_sigstore_pb.py` interface which is mostly the
same as `sign_sigstore.py` except that it uses the sigstore protobuf
specs instead of using `sigstore-python`. Since `sigstore-python` only
assumes signing with Sigstore, we have to do this.

Then, we add abilities to sign with key pairs and with certificates.

This resolves several bugs that existed in the `signature/`
implementations, making sure to properly check everything and to build a
PAE that is correctly build (proper capitalization, use bytes, by type
correct).

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>