Merge `signing` and `signature` and fix multiple bugs with the traditional signing support (#413)
Migrate some verification code to be the same in all paths. Use the correct PAE encoding (lowercase `v`, properly use bytes, don't mix `str` and `bytes`, be explicit in the encoding). Verify all the fields in Sigstore bundle. Unify the API for the signers and verifiers, all now can work with `pathlib` objects instead of some using `str`.

* Simplify: `sigstore.Sigstore<.>` to `sigstore.<.>`

Also removed docstrings for methods that are overrides with no actual
additional content, and cleaned up some docstring.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Extract payload type checking to `signing.Verifier`

It was just in `sign_sigstore.Verifier` but we actually need to perform
these checks each time we check a signature, so we move them to the main
class and have subclasses override a helper function instead.

I also added the remaining changes to the docstrings here.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Use `ValueError` instead of custom error.

In all the codebase, except what gets rewritten now, we use Python
errors. A custom error hierarchy would make sense only after we fuzzed
the entire codebase to determine all possible errors and sort them into
a proper ontology.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* fixup on split verify

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Add traditional signing signers and verifyiers.

First, we add a new `sign_sigstore_pb.py` interface which is mostly the
same as `sign_sigstore.py` except that it uses the sigstore protobuf
specs instead of using `sigstore-python`. Since `sigstore-python` only
assumes signing with Sigstore, we have to do this.

Then, we add abilities to sign with key pairs and with certificates.

This resolves several bugs that existed in the `signature/`
implementations, making sure to properly check everything and to build a
PAE that is correctly build (proper capitalization, use bytes, by type
correct).

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Switch to using the newly created interface

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Remove dead code

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Fix a lingering type error

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Also remove the now unused imports

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>