Use the content hasher name for serialization type
Trying to guess serialization format when it uses shard results in

```
ValueError                                Traceback (most recent call last)
<ipython-input-43-18521b474320> in <cell line: 0>()
      1 for model in all_models:
----> 2     verification_config.verify(model, f"{model}_sharded.sig")

5 frames
/usr/local/lib/python3.11/dist-packages/model_signing/hashing.py in _build_stream_hasher(self, hashing_algorithm)
    137             return memory.BLAKE2()
    138
--> 139         raise ValueError(f"Unsupported hashing method {hashing_algorithm}")
    140
    141     def _build_file_hasher_factory(

ValueError: Unsupported hashing method sha256-sharded-1000000000
```

We should use the internal hash name, not the mangled one (so it would
be `sha256`).

We'll need another patch release after this for the colab, but this
should be the last issue.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>