cli: Take identity parameters from certificate if not passed
Make the --identity and --identity_provider command line parameters
optional and if either one of them is not given on the command line then
try to take them from the Fulcio certificate found in the signature file.
Ask for confirmation from the user in this case.

Introduce a --yes option for 'verify sigstore' for automatic confirmation
of any identity taken from the Fulcion certificate. This allows a user to
just do a integrity check on the model.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>