Add compatibility layer with v0.2 signatures.
This is temporary but we need to ensure that signatures that were
generated using the v0.2 CLI can still be verified at v1.0 and later.

There are 2 issues that we needed to fix:

- the PAE encoding mixed `bytes` and `str` types resulting in additional
  escaping being added to the statement that gets signed. Verification
  must replicate the same mix-up to gurantee that the signed material
  can be verified.
- the in-toto predicate was different: v0.2 is not compatible with any
  model signature extensions, is not extensible and could be manipulated
  to be incomplete (not catching all possible alterations of a model).
  Fortunately, we used different predicate types, so it is easy to catch
  and add the support.

Note that this support is only added for verification size. Any new
signature must be generated using the v1.0 format.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>