Add a PR template. (#445)
This is to ensure that we have a checklist that reminds people to add to
`CHANGELOG.md` (added in #444), sign the commits, add docstrings, type
annotations and tests. Remind about updating documentation. Give on
point examples on when the changelog needs to be updated.

Previous template (global across all repos) used two sections that were
usually replaced with `NONE` or deleted. These are no longer here.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>