Enable signing with PKCS #11 devices (HSMs) (#411)
* Implement a class for handling of PKCS #11 URIs

Implement a class for handling of PKCS #11 URIs that will be used for
parsing the URIs and accessing its components. This will enable
signing with PKCS #11 enabled devices, such as HSMs.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* cli: Add support for signing with PKCS #11 URIs

Add support for signing with PKCS #11 keys that are described as URIs from
RFC 7512. Extend the existing signing with private key ('key' and
'certificate') to also accept PKCS #11 URIs, which are detected by their
prefix 'pkcs11:'.

Add documentation for PKCS #11 support.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* test: Add a tests for PKCS #11 with SoftHSM2 used for signing

Extend the python tests with a test case for pkcs11 using SoftHSM2 for
signing.

Add a script that runs the SoftHSM test.

Install softhsm2 and gnutls-bin on ubuntu-latest to be able to run the
python integration test.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* test: Skip softhsm setup if user already set it up

To speed up the integration test, skip running 'softhsm_setup setup' if
the key URI can be gotten from the softhsm_setup script.

Setup softhsm before running integration tests on github action.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>