signing: Fix handling of certificate that has no KeyUsage
If a certificate without KeyUsage is encountered then the following error
is displayed:

Verification failed with error: \
   No <class 'cryptography.x509.extensions.KeyUsage'> extension was found

The error may be mistaken for a broken python cryptography library even
though it is related to the certificate missing 'KeyUsage'. Fix this error
by intercepting the x509.ExtensionNotFound exception and trying with the
ExtendedKeyUsage next.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>