hashing: Fix paths of to-be-ignored files by using relpath
To avoid making an existing signature file part of the signature by
providing a path like the following, use relpath on each path given by the
user.

   ../curr-dir/model-dir/model.sig

In this case '../curr-dir' is basically a no-op and the path can be
rewritten (by relpath) like this:

   model-dir/model.sig

With this path the filter function should_ignore will properly ignore
the model.sig file.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>